% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_anonymiser.R
\name{anonymise_dataset}
\alias{anonymise_dataset}
\title{Function to anonymise dataset}
\usage{
anonymise_dataset(data, variables_to_remove = c())
}
\arguments{
\item{data}{the dataset you wish to anonymise}

\item{variables_to_remove}{the names of the sensitive variables you want to remove}
}
\value{
An anonymised dataset
}
\description{
The following function removes all the assigned sensitive informations from the raw dataset
}
\references{
\url{https://github.com/agualtieri/mcla_cleaneR}
}
\author{
Alberto Gualtieri, \email{alberto.gualtieri@reach-initiative.org}
}
\keyword{check,}
\keyword{mcla,}
\keyword{platypus}
\keyword{quality}
\keyword{yemen,}
