% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_log_to_long_format.R
\name{quality_checks_log_to_long_format}
\alias{quality_checks_log_to_long_format}
\title{Function to melt quality check log from long to wide}
\usage{
quality_checks_log_to_long_format(data, variable_name = "variable",
  value_name = "value", meta_not_to_transform = c())
}
\arguments{
\item{data}{the dataset you wish to transform from wide to long}

\item{variable_name}{the name of the measured variable. The default name is 'variable'}

\item{value_name}{name of the transformed data value. The default name is 'value'}

\item{meta_not_to_transform}{a list of metadata you wish to keep as id variables.}
}
\value{
A transformed quality check log
}
\description{
The following function turns the quality check output from wide to long format.
}
\references{
\url{https://github.com/agualtieri/dataqualitycontrol}
}
\author{
Alberto Gualtieri, \email{alberto.gualtieri@reach-initiative.org}
}
\keyword{check,}
\keyword{mcla,}
\keyword{platypus}
\keyword{quality}
\keyword{yemen,}
