% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_on_multiple.R
\name{separate_on_multiple}
\alias{separate_on_multiple}
\title{Function to separate a string variable into to columns using multiple separators}
\usage{
separate_on_multiple(data, var_to_separate, sep1, sep2 = NULL,
  sep3 = NULL)
}
\arguments{
\item{data}{the dataset containing the variable that needs to be separated}

\item{var_to_separate}{the variable that needs to be separated}

\item{sep1}{the first separator included in the variable that needs to be separated}

\item{sep2}{the second separator included in the variable that needs to be separated (default is NULL)}

\item{sep3}{the third separator included in the variable that needs to be separated (defaul is NULL)}
}
\value{
the @param var_to_separate will be divided into two column based on the included separators
}
\description{
The following function helps you in separating a variable into two column using up to three separators
}
\references{
\url{https://github.com/agualtieri/dataqualitycontrol}
}
\author{
Alberto Gualtieri, \email{alberto.gualtieri@reach-initiative.org}
}
\keyword{check,}
\keyword{mcla,}
\keyword{platypus}
\keyword{quality}
\keyword{yemen,}
