% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_quality_check.R
\name{batch_issue_checks}
\alias{batch_issue_checks}
\title{Batch check of predefined data quality conditions}
\usage{
batch_issue_checks(df, conditions, tests, meta_to_keep = c())
}
\arguments{
\item{df}{the dataframe you wish to check for invalid combinations of variables}

\item{conditions}{the set of conditions you wish to check}

\item{tests}{the set of test related to each condition}

\item{meta_to_keep}{a list of metadata you wish to include in the final output}
}
\value{
Standardized issues table for quality check in long format.
}
\description{
The following function checks for the presence of invalid combinations of variables within the
supplied data and produced and editable output that can be used for data quality review.
}
\references{
\url{https://github.com/agualtieri/dataqualitycontrol}
}
\author{
Alberto Gualtieri, \email{alberto.gualtieri@reach-initiative.org}
}
\keyword{check,}
\keyword{mcla,}
\keyword{platypus}
\keyword{quality}
\keyword{yemen,}
