% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataqualitycontrol.R
\docType{package}
\name{dataqualitycontrol}
\alias{dataqualitycontrol}
\alias{dataqualitycontrol-package}
\title{dataqualitycontrol: A package to produce standardized data quality output}
\description{
The data quality control package provides four functions:
\itemize{
\item batch quality checks
\item quality log to long format
\item dataframe anonymiser
\item separate on multiple separators
}
}
